% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.R
\name{create_circle_polygon_residuals}
\alias{create_circle_polygon_residuals}
\title{Create Circle to Polygon Residual Family}
\usage{
create_circle_polygon_residuals(target_shape, params)
}
\description{
Create Circle to Polygon Residual Family
}
\keyword{internal}
