% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.R
\name{create_sphere_polyhedron_residuals}
\alias{create_sphere_polyhedron_residuals}
\title{Create Sphere to Polyhedron Residual Family}
\usage{
create_sphere_polyhedron_residuals(target_shape, params)
}
\description{
Create Sphere to Polyhedron Residual Family
}
\keyword{internal}
