% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.R
\name{draw_polygon_standard}
\alias{draw_polygon_standard}
\title{Draw Polygon with Standardized Orientation}
\usage{
draw_polygon_standard(cx, cy, side_length, n, col = "black", lwd = 1)
}
\description{
Draw Polygon with Standardized Orientation
}
\keyword{internal}
