% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.R
\name{draw_regular_polygon}
\alias{draw_regular_polygon}
\title{Draw Regular Polygon (Fixed Orientation)}
\usage{
draw_regular_polygon(cx, cy, side_length, n, col = "black", lwd = 1)
}
\description{
Draw Regular Polygon (Fixed Orientation)
}
\keyword{internal}
