% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.R
\name{draw_square}
\alias{draw_square}
\title{Draw Square (Fixed Orientation)}
\usage{
draw_square(cx, cy, side_length, col = "black", lwd = 1)
}
\description{
Draw Square (Fixed Orientation)
}
\keyword{internal}
