% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_koch_deformity.R
\name{print.koch_deformity}
\alias{print.koch_deformity}
\title{Print Method for Koch Deformity Objects}
\usage{
\method{print}{koch_deformity}(x, ...)
}
\arguments{
\item{x}{Object of class "koch_deformity"}

\item{...}{Additional arguments (ignored)}
}
\description{
S3 print method for objects of class "koch_deformity"
}
