% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_koch_deformity.R
\name{summary.koch_deformity}
\alias{summary.koch_deformity}
\title{Summary Method for Koch Deformity Objects}
\usage{
\method{summary}{koch_deformity}(object, ...)
}
\arguments{
\item{object}{Object of class "koch_deformity"}

\item{...}{Additional arguments (ignored)}
}
\description{
S3 summary method for objects of class "koch_deformity"
}
